useUserList 훅으로 분리
